chore: rename patch to Unlock RevenueCat Pro
